Rename code/e2e-tests to e2e-sandbox
Clarifies that Playwright specs under e2e-sandbox target generated sandboxes, distinct from e2e-internal for the internal Storybook UI.
