+++ CODEOWNERS
@@ -62,4 +62,4 @@
 # /code/renderers/web-components/ @kasperpeulen @JReinhold
 
 # E2E
-# /code/e2e-tests/ @yannbf @valentinpalkovic
+# /code/e2e-sandbox/ @yannbf @valentinpalkovic
