+++ CONTRIBUTING.md
@@ -90,7 +90,7 @@ Here's a highlight of notable directories and files:
 │   ├── builders
 │   ├── chromatic.config.json
 │   ├── core                     # Core package for UI and API of Storybook
-│   ├── e2e-tests
+│   ├── e2e-sandbox
 │   ├── frameworks               # Different framework-bundler versions of Storybook
 │   ├── lib                      # CLI and plugins
 │   ├── node_modules
