+++ code/e2e-sandbox/addon-backgrounds.spec.ts
