+++ code/e2e-sandbox/addon-controls.spec.ts
