+++ code/e2e-sandbox/addon-onboarding.spec.ts
