+++ code/e2e-sandbox/addon-toolbars.spec.ts
