+++ code/e2e-sandbox/framework-nextjs.spec.ts
