+++ code/e2e-sandbox/framework-vue3.spec.ts
